make --extra-prog-path a part of shared project config
Motivation for this change comes from the desire to have
--extra-prog-path available when searching for VCS programs (at that
point, it turns, we only have a handle on a shared project config), but
it makes sense in general too.

Technically, the legacy flag configProgramPathExtra, which holds the
value of --extra-prog-path, did not propagate to shared project config
before, although the latter has a field for that sort of
thing (projectConfigProgPathExtra). So, we fix it. A little awkward is
that another legacy flag, globalProgPathExtra, also holds the same kind
of information, so, when turning legacy flags into a shared project
config we merge the two legacy flags and assign the result to
projectConfigProgPathExtra.