GHC 9.14 compat: replace async with text in a test for haddock-project
the test was using async from hackage as an exmaple global package.
Currently, async has a too restrictive upper bound on base. It'll be
more robust to rely on a boot package instead. Ideally we'd not rely on
anything from Hackage in tests, but that's an issue for another day.